[COMBINE] 머지 - create issue form 머지 - 오타 수정
