+++ frontEnd/src/models/SidebarModel.js
@@ -1,4 +1,6 @@
 import React, {createContext, useState, useEffect, useReducer} from 'react';
+import axiosMaker from '~/*/utils/axios/axiosMaker';
+
 export const SidebarModelContext = createContext();
 
 const SidebarModelConsumer = ({children}) => {
@@ -6,7 +8,7 @@ const SidebarModelConsumer = ({children}) => {
   const [milestone, setMilestone] = useState();
   const [assignees, setAssignees] = useState([]);
 
-  const onUpdateLables = (label) => {
+  const onUpdateLabels = (label) => {
     if (labels.includes(label)) {
       // 찾고, splice
       return;
@@ -37,11 +39,29 @@ const SidebarModelConsumer = ({children}) => {
     setIssueContent(e.target.value);
   };
 
+  
+  const requestImageUpload = async (formData) => {
+    const axiosInstance = axiosMaker();
+    const result = await axiosInstance.post('/api/issue/image', formData, {
+      headers: {
+        'content-type': 'multipart/form-data',
+      },
+    });
+
+    if (result.status === 200) {
+      setIssueContent((content) => content + result.data.filename);
+    }
+  };
+
+  const requests = {
+    requestImageUpload,
+  };
+
   return (
     <SidebarModelContext.Provider
       value={{
         labels,
-        onUpdateLables,
+        onUpdateLabels,
         milestone,
         onUpdateMilestone,
         assignees,
@@ -50,6 +70,7 @@ const SidebarModelConsumer = ({children}) => {
         onUpdateIssueTitle,
         issueContent,
         onUpdateIssueContent,
+        requests
       }}
     >
       {children}
